Revamped HiveThriftServer2Suite for robustness
